Stop using `get_settings` during import time
