+++ src/napari/layers/image/image.py
@@ -12,10 +12,12 @@ from scipy import ndimage as ndi
 
 from napari.layers._data_protocols import LayerDataProtocol
 from napari.layers._multiscale_data import MultiScaleData
+from napari.layers._scalar_field._scalar_field_constants import (
+    ImageProjectionMode,
+)
 from napari.layers._scalar_field._slice import _ScalarFieldSliceResponse
 from napari.layers._scalar_field.scalar_field import ScalarFieldBase
 from napari.layers.image._image_constants import (
-    ImageProjectionMode,
     ImageRendering,
     Interpolation,
     InterpolationStr,
