+++ src/napari/layers/points/_points_key_bindings.py
@@ -15,7 +15,9 @@ from napari.utils.interactions import Shortcut
 from napari.utils.notifications import show_info
 from napari.utils.translations import trans
 
-shortcuts = get_settings().shortcuts.shortcuts
+
+def shortcuts() -> dict[str, list[str]]:
+    return get_settings().shortcuts.shortcuts
 
 
 def register_points_action(
@@ -83,7 +85,7 @@ def select_all_in_slice(layer: Points) -> None:
         show_info(
             trans._(
                 'Deselected all points in this slice, use {shortcut} to select/deselect all points on the layer. ({n_total} selected)',
-                shortcut=Shortcut(shortcuts['napari:select_all_data'][0]),
+                shortcut=Shortcut(shortcuts()['napari:select_all_data'][0]),
                 n_total=len(layer.selected_data),
                 deferred=True,
             )
@@ -97,7 +99,7 @@ def select_all_in_slice(layer: Points) -> None:
                 'Selected {n_new} points in this slice only, use {shortcut} to append to existing selection. ({n_total} selected)',
                 n_new=len(new_selected),
                 shortcut=Shortcut(
-                    shortcuts['napari:select_append_all_in_slice'][0]
+                    shortcuts()['napari:select_append_all_in_slice'][0]
                 ),
                 n_total=len(layer.selected_data),
                 deferred=True,
@@ -118,7 +120,7 @@ def select_append_all_in_slice(layer: Points) -> None:
         show_info(
             trans._(
                 'Deselected all points in this slice, use {shortcut} to select/deselect all points on the layer. ({n_total} selected)',
-                shortcut=Shortcut(shortcuts['napari:select_all_data'][0]),
+                shortcut=Shortcut(shortcuts()['napari:select_all_data'][0]),
                 n_total=len(layer.selected_data),
                 deferred=True,
             )
@@ -131,7 +133,7 @@ def select_append_all_in_slice(layer: Points) -> None:
             trans._(
                 'Appended {n_new} points in this slice to the selection, use {shortcut} to select all points on the layer. ({n_total} selected)',
                 n_new=len(new_selected),
-                shortcut=Shortcut(shortcuts['napari:select_all_data'][0]),
+                shortcut=Shortcut(shortcuts()['napari:select_all_data'][0]),
                 n_total=len(layer.selected_data),
                 deferred=True,
             )
